+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -19,7 +19,6 @@ import org.junit.jupiter.api.io.TempDir;
 
 import seedu.address.commons.core.index.UniqueId;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.listcommands.ListCommand;
 import seedu.address.logic.commands.addcommands.AddBuyerCommand;
 import seedu.address.logic.commands.addcommands.AddDelivererCommand;
 import seedu.address.logic.commands.addcommands.AddSupplierCommand;
