+++ src/test/java/seedu/address/logic/commands/CheckCommandTest.java
@@ -2,7 +2,6 @@ package seedu.address.logic.commands;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.commands.checkcommands.CheckCommand.MESSAGE_SUCCESS;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST;
@@ -11,7 +10,11 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_SECOND;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.checkcommands.CheckBuyerCommand;
 import seedu.address.logic.commands.checkcommands.CheckCommand;
+//import seedu.address.logic.commands.checkcommands.CheckOrderCommand;
+//import seedu.address.logic.commands.checkcommands.CheckPetCommand;
+import seedu.address.logic.commands.checkcommands.CheckSupplierCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
@@ -34,15 +37,15 @@ public class CheckCommandTest {
 
     @Test
     public void equals() {
-        CheckCommand firstCommand = new CheckCommand("BUYER", INDEX_FIRST);
-        CheckCommand secondCommand = new CheckCommand("BUYER", INDEX_SECOND);
-        CheckCommand thirdCommand = new CheckCommand("SUPPLIER", INDEX_FIRST);
+        CheckCommand firstCommand = new CheckBuyerCommand(INDEX_FIRST);
+        CheckCommand secondCommand = new CheckBuyerCommand(INDEX_SECOND);
+        CheckCommand thirdCommand = new CheckSupplierCommand(INDEX_FIRST);
 
         //same object -> returns true
         assertTrue(firstCommand.equals(firstCommand));
 
         //same values -> returns true
-        CheckCommand firstCommandCopy = new CheckCommand("BUYER", INDEX_FIRST);
+        CheckCommand firstCommandCopy = new CheckBuyerCommand(INDEX_FIRST);
         assertTrue(firstCommand.equals(firstCommandCopy));
 
         // different types -> returns false
@@ -60,46 +63,36 @@ public class CheckCommandTest {
 
     @Test
     public void execute_invalidIndex_throwsCommandException() {
-        CheckCommand command = new CheckCommand("BUYER", Index.fromOneBased(100));
+        CheckCommand command = new CheckBuyerCommand(Index.fromOneBased(100));
         assertThrows(CommandException.class, () -> command.execute(bModel));
     }
 
     @Test
     public void execute() {
         String checkType = "BUYER";
-        CommandResult expected = CommandResult.createCheckCommandResult(String.format(MESSAGE_SUCCESS, checkType,
-                        INDEX_FIRST.getOneBased()), checkType, INDEX_FIRST);
-        CheckCommand command = new CheckCommand(checkType, INDEX_FIRST);
+        CommandResult expected = new CommandResult(String.format(CheckBuyerCommand.MESSAGE_SUCCESS,
+                        INDEX_FIRST.getOneBased()));
+        CheckCommand command = new CheckBuyerCommand(INDEX_FIRST);
 
         // buyer -> success
         assertCommandSuccess(command, bModel, expected, bExpectedModel);
 
         // supplier -> success
-        checkType = "SUPPLIER";
-        expected = CommandResult.createCheckCommandResult(String.format(MESSAGE_SUCCESS, checkType,
-                INDEX_FIRST.getOneBased()), checkType, INDEX_FIRST);
-        command = new CheckCommand(checkType, INDEX_FIRST);
+        expected = new CommandResult(String.format(CheckSupplierCommand.MESSAGE_SUCCESS,
+                INDEX_FIRST.getOneBased()));
+        command = new CheckSupplierCommand(INDEX_FIRST);
         assertCommandSuccess(command, sModel, expected, sExpectedModel);
 
         // order -> success
-        checkType = "ORDER";
-        expected = CommandResult.createCheckCommandResult(String.format(MESSAGE_SUCCESS, checkType,
-                INDEX_FIRST.getOneBased()), checkType, INDEX_FIRST);
-        command = new CheckCommand(checkType, INDEX_FIRST);
-        assertCommandSuccess(command, oModel, expected, oExpectedModel);
+        //expected = new CommandResult(String.format(CheckOrderCommand.MESSAGE_SUCCESS,
+        //        INDEX_FIRST.getOneBased()));
+        //command = new CheckOrderCommand(INDEX_FIRST);
+        //assertCommandSuccess(command, oModel, expected, oExpectedModel);
 
         // pet -> success
-        checkType = "PET";
-        expected = CommandResult.createCheckCommandResult(String.format(MESSAGE_SUCCESS, checkType,
-                INDEX_FIRST.getOneBased()), checkType, INDEX_FIRST);
-        command = new CheckCommand(checkType, INDEX_FIRST);
-        assertCommandSuccess(command, pModel, expected, pExpectedModel);
-
-        // invalid check type -> do nothing
-        checkType = "SOME CHECK TYPE";
-        expected = CommandResult.createCheckCommandResult(String.format(MESSAGE_SUCCESS, checkType,
-                INDEX_FIRST.getOneBased()), checkType, INDEX_FIRST);
-        command = new CheckCommand(checkType, INDEX_FIRST);
-        assertCommandSuccess(command, pModel, expected, pExpectedModel);
+        //expected = new CommandResult(String.format(CheckPetCommand.MESSAGE_SUCCESS,
+        //        INDEX_FIRST.getOneBased()));
+        //command = new CheckPetCommand(INDEX_FIRST);
+        //assertCommandSuccess(command, pModel, expected, pExpectedModel);
     }
 }
