+++ src/test/java/seedu/address/logic/commands/editcommands/EditDelivererCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.editcommands;
+/*package seedu.address.logic.commands.editcommands;
 
 import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -108,6 +108,7 @@ public class EditDelivererCommandTest {
     /**
      * A default model stub that have all the methods failing.
      */
+/*
     private class ModelStub implements Model {
         @Override
         public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
@@ -414,6 +415,7 @@ public class EditDelivererCommandTest {
     /**
      * A Model stub that contains a single buyer.
      */
+/*
     private class ModelStubWithDeliverer extends ModelStub {
         private final Deliverer deliverer;
 
@@ -432,6 +434,7 @@ public class EditDelivererCommandTest {
     /**
      * A Model stub that always accept the buyer being added.
      */
+/*
     private class ModelStubAcceptingDelivererAdded extends ModelStub {
         private ArrayList<Deliverer> deliverers = new ArrayList<>();
 
@@ -471,3 +474,4 @@ public class EditDelivererCommandTest {
         }
     }
 }
+*/
