+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -21,21 +21,18 @@ import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.checkcommands.CheckBuyerCommand;
-import seedu.address.logic.commands.checkcommands.CheckCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.checkcommands.CheckOrderCommand;
-import seedu.address.logic.commands.checkcommands.CheckSupplierCommand;
-import seedu.address.logic.commands.listcommands.ListAllCommand;
-import seedu.address.logic.commands.listcommands.ListCommand;
 import seedu.address.logic.commands.MatchCommand;
 import seedu.address.logic.commands.addcommands.AddBuyerCommand;
 import seedu.address.logic.commands.addcommands.AddDelivererCommand;
-//import seedu.address.logic.commands.addcommands.AddPetCommand;
 import seedu.address.logic.commands.addcommands.AddPetCommand;
 import seedu.address.logic.commands.addcommands.AddSupplierCommand;
+import seedu.address.logic.commands.checkcommands.CheckBuyerCommand;
+import seedu.address.logic.commands.checkcommands.CheckCommand;
+import seedu.address.logic.commands.checkcommands.CheckOrderCommand;
+import seedu.address.logic.commands.checkcommands.CheckSupplierCommand;
 import seedu.address.logic.commands.deletecommands.DeleteBuyerCommand;
 import seedu.address.logic.commands.deletecommands.DeleteDelivererCommand;
 import seedu.address.logic.commands.deletecommands.DeleteSupplierCommand;
@@ -46,6 +43,8 @@ import seedu.address.logic.commands.editcommands.EditSupplierCommand;
 import seedu.address.logic.commands.filtercommands.FilterOrderCommand;
 import seedu.address.logic.commands.filtercommands.FilterPetCommand;
 import seedu.address.logic.commands.findcommands.FindCommand;
+import seedu.address.logic.commands.listcommands.ListAllCommand;
+import seedu.address.logic.commands.listcommands.ListCommand;
 import seedu.address.logic.commands.sortcommands.SortCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.logic.parser.findcommandparser.FindBuyerCommandParser;
